test(calendar): extract shared setViewportWidth helper
